Added first draft of PineconeNeo4jRetriever
