+++ src/neo4j_genai/retrievers/external/utils.py
@@ -13,6 +13,22 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from .pinecone import PineconeNeo4jRetriever
+from typing import Optional
 
-__all__ = ["PineconeNeo4jRetriever"]
+from neo4j_genai.neo4j_queries import get_query_tail
+
+
+def get_match_query(
+    return_properties: Optional[list[str]] = None, retrieval_query: Optional[str] = None
+) -> str:
+    match_query = (
+        "UNWIND $match_params AS match_param "
+        "WITH match_param[0] AS match_id_value, match_param[1] AS score "
+        "MATCH (node) "
+        "WHERE node[$id_property] = match_id_value "
+    )
+    return match_query + get_query_tail(
+        return_properties=return_properties,
+        retrieval_query=retrieval_query,
+        fallback_return="RETURN node, score",
+    )
