+++ src/neo4j_genai/retrievers/external/weaviate/__init__.py
@@ -13,6 +13,6 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from .weaviate import WeaviateNeo4jRetriever, get_match_query
+from .weaviate import WeaviateNeo4jRetriever
 
-__all__ = ["WeaviateNeo4jRetriever", "get_match_query"]
+__all__ = ["WeaviateNeo4jRetriever"]
